EigenSDK and AVSs in Rust (#5)
* Changes from moving repo

* Debugging and some Fixes

* Fixes all contract errors and submodule issue

* Add simple README

* More registration and signing changes

* Registration WIP

* Moving back to Rust Contract Deployment

* Contracts Deployment WIP

* Testing utilities WIP and reorganization

* Debugging Contracts and some cleanup

* Build Scripts and register debugging

* Small build script fix

* Debugging, Testing, and Cleanup

* Utilities for Testing, Contracts, and Operators

* Workflow, testing readme, and build script

* BLS Pairing Debugging and Fix

* Operator metadata

* Metadata format fix

* Metadata description

* Foundry workflow fix

* Contract Debugging for finishing Operator up

* Registration and Aggregation Finalization

* Workflow Fix

* Debugging

* Debugging

* Debugging Workflow

* Fix GH workflow

* Debugging Forge steps

* Debugging Forge steps

* Build Script Fix

* Workflow WIP

* Workflow WIP

* Fmt and Fix

* Clippy workflow debugging

* Contract path Debugging

* Contract path Debugging

* Recursively checkout submodules

* Contract Building WIP

* Update Tests

* Add tests to CI

* Add forge to workflow testing job

* Run build scripts before tests

* README and Licenses

---------

+++ utils/src/avs_registry/reader.rs
@@ -337,7 +337,7 @@ impl<T: Config> AvsRegistryChainReaderTrait for AvsRegistryContractManager<T> {
         let start = start_block;
         let mut end = stop_block;
         if start_block == 0 && stop_block == 0 {
-            end = self.eth_client_http.get_block_number().await? as u64;
+            end = self.eth_client_http.get_block_number().await?;
         }
 
         for i in (start..=end).step_by(block_range as usize) {
